fix: Incorrect incremental channel system
Fix & Refactor all the task due to a wrong usage.
Tests successfull on personal Discord server.
Some clean in the import (one line)

__Basic Usage :__
3 lobby channel id to define & 1 timer
When a user enter in one of lobby, dedicated channel following the lobby name is created & incremented
When a channel is empty, the channel is deleted

__Restriction : __
Mentoring channel creation only with a mentor role